+++ tests/e2etests/test_school_contact_param.py
@@ -20,13 +20,14 @@ from setup_e2e import get_fig_dir
 
 pars = dict(
     n                       = 15e3,
-    rand_seed               = 1,
+    rand_seed               = 123,
     with_non_teaching_staff = 1
 )
 
 
+@pytest.mark.parametrize("school_mixing_type", ['random', 'age_clustered', 'age_and_class_clustered'])
 @pytest.mark.parametrize("average_class_size", [10, 50])
-def test_average_class_size(average_class_size, do_show, do_save, get_fig_dir, quantiles=None):
+def test_average_class_size(average_class_size, school_mixing_type, do_show, do_save, get_fig_dir, quantiles=None):
     """
     Test case to check average_class_size by taking average of student-student contacts
 
@@ -36,30 +37,45 @@ def test_average_class_size(average_class_size, do_show, do_save, get_fig_dir, q
     Returns:
         None
     """
+    sp.logger.info(f"Test average_class_size: {average_class_size}.")
     testpars = dict(
-        average_class_size = average_class_size,
-        # average_student_teacher_ratio = average_class_size,  # DM: note that this parameter will overide the average class size parameter when school mixing types are something other than random or undefined (which defaults to random) --- method refactor work for schools will clarify these relationships
+        average_class_size            = average_class_size,
+        average_student_teacher_ratio = 20,  # DM: note that this parameter will overide the average class size parameter when school mixing types are something other than random or undefined (which defaults to random) --- method refactor work for schools will clarify these relationships
+        with_school_types             = 1,
+        school_mixing_type            = school_mixing_type,
     )
     pop = sp.Pop(**pars, **testpars)
     plotting_kwargs = sc.objdict(do_show=do_show, do_save=do_save, figdir=get_fig_dir)
+
     contacts = sp.get_contact_counts_by_layer(pop.popdict, with_layer_ids=1)[0]
-    plotting_kwargs.append("title_prefix", f"Average Class Size = {average_class_size}")
-    plotting_kwargs.append("figname", f"contact_average_class_size_{average_class_size}")
+    title = f"Average Class Size = {average_class_size} {pop.school_pars.school_mixing_type}"
+    plotting_kwargs.append("title_prefix", title)
+    plotting_kwargs.append("figname", f"contact_average_class_size_{average_class_size}_{pop.school_pars.school_mixing_type}")
     sp.plot_contact_counts(contacts, **plotting_kwargs)
+
     counts = []
     if not pop.school_pars.with_school_types:
         counts.extend(contacts['sc_student']['all'])
         counts.extend(contacts['sc_teacher']['all'])
         counts.extend(contacts['sc_staff']['all'])
 
-    elif pop.school_pars.with_school_types and pop.school_pars.school_mixing_type == 'age_and_class_clustered':
+    elif (pop.school_pars.with_school_types) & (pop.school_pars.school_mixing_type == 'random'):
+        counts.extend(contacts['sc_student']['all'])
+        counts.extend(contacts['sc_teacher']['all'])
 
+    elif (pop.school_pars.with_school_types) & (pop.school_pars.school_mixing_type == 'age_clustered'):
         counts.extend(contacts['sc_student']['sc_student'])
+
+    elif (pop.school_pars.with_school_types) & (pop.school_pars.school_mixing_type == 'age_and_class_clustered'):
+        counts.extend(contacts['sc_student']['sc_student'])
+        if pop.school_pars.average_class_size < pop.school_pars.average_student_teacher_ratio:
+            average_class_size = pop.school_pars.average_student_teacher_ratio
 
     sp.check_poisson(actual=counts, expected=average_class_size, label='average_class_size', check='dist')
     # visual check with scipy.stats.probplot -- temporary, just to show that the null hypothesis should pass here for the distribution
     fig, ax = plt.subplots(1, 1)
     res = stats.probplot(counts, dist=stats.poisson, sparams=(average_class_size, ), plot=ax)
+    ax.set_title(f'Probability Plot for {title}')
     if do_show:
         plt.show()
     plt.close()
@@ -225,4 +241,11 @@ def get_teacher_staff_ratio(popdict, varname, varvalue, do_show, do_save, fig_di
 
 
 if __name__ == "__main__":
-    pytest.main(['-vs', __file__])
+    # pytest.main(['-vs', __file__])
+
+    # # run an example in main and plot!
+    average_class_size = 30
+    school_mixing_type = 'random'
+    do_show = 1
+    do_save = 0
+    test_average_class_size(average_class_size, school_mixing_type, do_show, do_save, get_fig_dir)